낚시 미니게임 노트 이탈 현상 해결 완료
- fish 씬에서 3d가 아닌 2d로 설정해서 실행하시기 바랍니다. 노트의 충돌 판정이 2d box collider입니다.

+++ Assets/InGame/02.Scripts/Manager/InputManager.cs
@@ -98,7 +98,7 @@ public class InputManager : MonoBehaviour
 
     private void CheckInput()
     {
-        if(InGameUIManager.UICheck == (int)InGameUIManager.UI_BitFlag.SHOP)
+        if(InGameUIManager.UICheck == InGameUIManager.UI_BitFlag.SHOP)
         {
             if(nowState != InputState.NULL)
             {
